requirements.txt
anthropic==0.52.0
python-dotenv==1.1.0
pytest==8.3.4
pytest-asyncio==0.25.3
mcp==1.26.0
